Move prewitt and sobel
