+++ src/filters/mod.rs
@@ -4,8 +4,6 @@ pub mod con_corr;
 pub mod gaussian;
 pub mod median;
 pub mod min_max;
-pub mod prewitt;
-pub mod sobel;
 pub mod symmetry;
 
 // TODO We might want to offer all NumPy mode (use PadMode instead)
