+++ src/lib.rs
@@ -12,15 +12,13 @@ mod morphology;
 mod pad;
 
 pub use filters::{
-    con_corr::{convolve, convolve1d, correlate, correlate1d},
+    con_corr::{convolve, convolve1d, correlate, correlate1d, prewitt, sobel},
     gaussian::{gaussian_filter, gaussian_filter1d},
     median::median_filter,
     min_max::{
         maximum_filter, maximum_filter1d, maximum_filter1d_to, minimum_filter, minimum_filter1d,
         minimum_filter1d_to,
     },
-    prewitt::prewitt,
-    sobel::sobel,
     BorderMode,
 };
 pub use interpolation::{spline_filter, spline_filter1d};
